Some general fixes and cleanup

+++ SOD.Common/BepInEx/PluginController.cs
@@ -160,13 +160,14 @@ namespace SOD.Common.BepInEx
         {
             try
             {
-                if (_createdNewConfigFile || string.IsNullOrWhiteSpace(ConfigFile.ConfigFilePath) || !File.Exists(ConfigFile.ConfigFilePath)) return;
+                if (_createdNewConfigFile || ConfigFile == null || string.IsNullOrWhiteSpace(ConfigFile.ConfigFilePath) || !File.Exists(ConfigFile.ConfigFilePath)) return;
                 if (!HasConfigurationBindings())
                 {
                     File.Delete(ConfigFile.ConfigFilePath);
                     Plugin.Log.LogInfo($"Deleted outdated configuration file for mod \"{PluginGUID}\".");
                     return;
                 }
+                if (Config == null) return;
 
                 var helper = new ConfigHelper(ConfigFile.ConfigFilePath);
                 var configEntries = helper.GetConfigEntries();
@@ -177,7 +178,7 @@ namespace SOD.Common.BepInEx
                     .Where(a => a.GetCustomAttribute<BindingAttribute>() != null)
                     .Select(a =>
                     {
-                        var (section, key) = ConfigHelper.SplitIdentifier(a.GetCustomAttribute<BindingAttribute>().Name);
+                        var (section, key) = ConfigHelper.SplitIdentifier(a.GetCustomAttribute<BindingAttribute>().Name ?? "");
                         return new { Section = section, Key = key };
                     })
                     .GroupBy(a => a.Section)
@@ -189,7 +190,7 @@ namespace SOD.Common.BepInEx
                     var section = configEntry.Key;
                     if (properties.TryGetValue(section, out var validValues))
                     {
-                        var invalidValues = configEntry.Value.Except(validValues);
+                        var invalidValues = configEntry.Value.Except(validValues ?? Array.Empty<string>());
                         foreach (var value in invalidValues)
                             helper.RemoveEntry(value, section);
                     }
